Split the classes RealmMigration into smaller classes
