+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/auth/db/migration/MigrateAuthTo003.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2022 The Matrix.org Foundation C.I.C.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.matrix.android.sdk.internal.auth.db.migration
+
+import io.realm.DynamicRealm
+import org.matrix.android.sdk.api.auth.data.Credentials
+import org.matrix.android.sdk.api.auth.data.sessionId
+import org.matrix.android.sdk.internal.auth.db.SessionParamsEntityFields
+import org.matrix.android.sdk.internal.di.MoshiProvider
+import org.matrix.android.sdk.internal.util.database.RealmMigrator
+import timber.log.Timber
+
+class MigrateAuthTo003(realm: DynamicRealm) : RealmMigrator(realm, 3) {
+
+    override fun doMigrate(realm: DynamicRealm) {
+        Timber.d("Update SessionParamsEntity primary key, to allow several sessions with the same userId")
+
+        realm.schema.get("SessionParamsEntity")
+                ?.removePrimaryKey()
+                ?.addField(SessionParamsEntityFields.SESSION_ID, String::class.java)
+                ?.setRequired(SessionParamsEntityFields.SESSION_ID, true)
+                ?.transform {
+                    val credentialsJson = it.getString(SessionParamsEntityFields.CREDENTIALS_JSON)
+
+                    val credentials = MoshiProvider.providesMoshi()
+                            .adapter(Credentials::class.java)
+                            .fromJson(credentialsJson)
+
+                    it.set(SessionParamsEntityFields.SESSION_ID, credentials!!.sessionId())
+                }
+                ?.addPrimaryKey(SessionParamsEntityFields.SESSION_ID)
+    }
+}
