+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/migration/MigrateSessionTo001.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2022 The Matrix.org Foundation C.I.C.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.matrix.android.sdk.internal.database.migration
+
+import io.realm.DynamicRealm
+import org.matrix.android.sdk.internal.database.model.RoomSummaryEntityFields
+import org.matrix.android.sdk.internal.util.database.RealmMigrator
+
+class MigrateSessionTo001(realm: DynamicRealm) : RealmMigrator(realm, 1) {
+
+    override fun doMigrate(realm: DynamicRealm) {
+        // Add hasFailedSending in RoomSummary and a small warning icon on room list
+        realm.schema.get("RoomSummaryEntity")
+                ?.addField(RoomSummaryEntityFields.HAS_FAILED_SENDING, Boolean::class.java)
+                ?.transform { obj ->
+                    obj.setBoolean(RoomSummaryEntityFields.HAS_FAILED_SENDING, false)
+                }
+    }
+}
