+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/migration/MigrateSessionTo002.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Matrix.org Foundation C.I.C.
+ * Copyright (c) 2022 The Matrix.org Foundation C.I.C.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.raw
+package org.matrix.android.sdk.internal.database.migration
 
 import io.realm.DynamicRealm
-import io.realm.RealmMigration
-import org.matrix.android.sdk.internal.raw.migration.MigrateGlobalTo001
-import timber.log.Timber
+import org.matrix.android.sdk.internal.util.database.RealmMigrator
 
-internal object GlobalRealmMigration : RealmMigration {
+class MigrateSessionTo002(realm: DynamicRealm) : RealmMigrator(realm, 2) {
 
-    // Current schema version
-    const val SCHEMA_VERSION = 1L
-
-    override fun migrate(realm: DynamicRealm, oldVersion: Long, newVersion: Long) {
-        Timber.d("Migrating Global Realm from $oldVersion to $newVersion")
-
-        if (oldVersion < 1) MigrateGlobalTo001(realm).perform()
+    override fun doMigrate(realm: DynamicRealm) {
+        realm.schema.get("HomeServerCapabilitiesEntity")
+                ?.addField("adminE2EByDefault", Boolean::class.java)
+                ?.transform { obj ->
+                    obj.setBoolean("adminE2EByDefault", true)
+                }
     }
 }
